+++ src/main/java/io/inprice/scrapper/api/rest/controller/ProductImportController.java
@@ -1,10 +1,18 @@
 package io.inprice.scrapper.api.rest.controller;
 
+import static spark.Spark.delete;
+import static spark.Spark.get;
+import static spark.Spark.post;
+
+import org.apache.commons.lang3.StringUtils;
+import org.eclipse.jetty.http.HttpStatus;
+
 import io.inprice.scrapper.api.framework.Beans;
 import io.inprice.scrapper.api.framework.Routing;
 import io.inprice.scrapper.api.helpers.Consts;
 import io.inprice.scrapper.api.helpers.Global;
 import io.inprice.scrapper.api.helpers.Responses;
+import io.inprice.scrapper.api.info.ServiceResponse;
 import io.inprice.scrapper.api.rest.component.Commons;
 import io.inprice.scrapper.api.rest.service.ProductCSVImportService;
 import io.inprice.scrapper.api.rest.service.ProductCodeImportService;
@@ -13,12 +21,8 @@ import io.inprice.scrapper.api.rest.service.ProductURLImportService;
 import io.inprice.scrapper.common.meta.ImportType;
 import io.inprice.scrapper.common.models.ImportProduct;
 import io.inprice.scrapper.common.utils.NumberUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.eclipse.jetty.http.HttpStatus;
 import spark.Request;
 import spark.Response;
-
-import static spark.Spark.*;
 
 public class ProductImportController {
 
@@ -32,36 +36,43 @@ public class ProductImportController {
 
         //find
         get(Consts.Paths.Product.IMPORT_BASE + "/:id", (req, res) -> {
+        	if (res.status() >= 400) return new ServiceResponse(res.status());
             return Commons.createResponse(res, importService.findById(NumberUtils.toLong(req.params(":id"))));
         }, Global.gson::toJson);
 
         //list
         get(Consts.Paths.Product.IMPORT_BASE + "s", (req, res) -> {
+        	if (res.status() >= 400) return new ServiceResponse(res.status());
             return Commons.createResponse(res, importService.getList());
         }, Global.gson::toJson);
 
         //delete
         delete(Consts.Paths.Product.IMPORT_BASE + "/:id", (req, res) -> {
+        	if (res.status() >= 400) return new ServiceResponse(res.status());
             return Commons.createResponse(res, importService.deleteById(NumberUtils.toLong(req.params(":id"))));
         }, Global.gson::toJson);
 
         //upload csv
         post(Consts.Paths.Product.IMPORT_CSV, "text/csv", (req, res) -> {
+        	if (res.status() >= 400) return new ServiceResponse(res.status());
             return createResponse(res, uploadCSV(req));
         }, Global.gson::toJson);
 
         //upload URL list
         post(Consts.Paths.Product.IMPORT_URL_LIST, "text/csv", (req, res) -> {
+        	if (res.status() >= 400) return new ServiceResponse(res.status());
             return createResponse(res, uploadURL(req));
         }, Global.gson::toJson);
 
         //upload ebay SKU list
         post(Consts.Paths.Product.IMPORT_EBAY_SKU_LIST, "text/plain", (req, res) -> {
+        	if (res.status() >= 400) return new ServiceResponse(res.status());
             return createResponse(res,  uploadCodeList(ImportType.EBAY_SKU, req));
         }, Global.gson::toJson);
 
         //upload amazon ASIN list
         post(Consts.Paths.Product.IMPORT_AMAZON_ASIN_LIST, "text/plain", (req, res) -> {
+        	if (res.status() >= 400) return new ServiceResponse(res.status());
             return createResponse(res,  uploadCodeList(ImportType.AMAZON_ASIN, req));
         }, Global.gson::toJson);
